Fix uint8_t channel must be interpreted as integral value for CLI11 and not as char which would lead to an ValidationError.

+++ src/net/rc/config/ConfigAddress.cpp
@@ -97,7 +97,7 @@ namespace net::rc::config {
         utils::PreserveErrno preserveErrno;
 
         channelOpt //
-            ->default_val(channel);
+            ->default_val(static_cast<int>(channel));
         Super::required(channelOpt, false);
     }
